Add add events into modules
