+++ week4/js-calenda-app/add-events/index.html
@@ -51,6 +51,6 @@
 
     <div id="modalBackDrop"></div>
 
-    <script src="./index.js"></script>
+    <script type="module" src="./main.mjs"></script>
   </body>
 </html>
